% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-through.R
\name{.autonomousDesign}
\alias{.autonomousDesign}
\title{Design matrix for the autonomous component}
\usage{
.autonomousDesign(design, group, n_samples)
}
\arguments{
\item{design}{NULL or a design matrix with one row per sample}

\item{group}{NULL or a factor with one value per sample}

\item{n_samples}{Number of samples}
}
\value{
Design matrix. An intercept-only design if both design and group
are NULL.
}
\description{
Design matrix for the autonomous component
}
\keyword{internal}
