% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.BUILD_FILES}
\alias{.BUILD_FILES}
\title{Files generated for each combination of species and exclude/min_len settings}
\usage{
.BUILD_FILES
}
\description{
Files generated for each combination of species and exclude/min_len settings
}
\keyword{internal}
