% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-through.R
\name{.chiBarSquare}
\alias{.chiBarSquare}
\title{P-values for a likelihood ratio test with parameters on the boundary}
\usage{
.chiBarSquare(stat, df)
}
\arguments{
\item{stat}{Likelihood ratio statistics}

\item{df}{Number of parameters tested}
}
\value{
p-values from a binomial mixture of chi-squared distributions with 0
to df degrees of freedom
}
\description{
P-values for a likelihood ratio test with parameters on the boundary
}
\keyword{internal}
