% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-through.R
\name{.crossWeighted}
\alias{.crossWeighted}
\title{Weighted cross products of a list of matrices for every row}
\usage{
.crossWeighted(mats, w)
}
\arguments{
\item{mats}{List of q matrices with the same dimensions}

\item{w}{Matrix of weights with the same dimensions}
}
\value{
Array with dimensions nrow x q x q where element \code{[i, u, v]} is
\code{sum_j w[i, j] * mats[[u]][i, j] * mats[[v]][i, j]}
}
\description{
Weighted cross products of a list of matrices for every row
}
\keyword{internal}
