% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.DEFAULT_EXCLUDE}
\alias{.DEFAULT_EXCLUDE}
\title{Default RepeatMasker classes excluded from the index}
\usage{
.DEFAULT_EXCLUDE
}
\description{
Default RepeatMasker classes excluded from the index
}
\keyword{internal}
