% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-through.R
\name{.downstreamPairs}
\alias{.downstreamPairs}
\title{Find TE loci downstream of the 3' end of same-strand hosts}
\usage{
.downstreamPairs(te, span, downstream)
}
\arguments{
\item{te}{GRanges of TE loci}

\item{span}{GRanges of host spans}

\item{downstream}{Maximum distance in bp downstream of the host 3' end}
}
\value{
data.frame with columns locus, host, position, and distance
}
\description{
Find TE loci downstream of the 3' end of same-strand hosts
}
\keyword{internal}
