% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-annotation.R
\name{.dupInfoToDT}
\alias{.dupInfoToDT}
\title{Extract duplicate info from json file as a data.table}
\usage{
.dupInfoToDT(jsonfile)
}
\arguments{
\item{jsonfile}{JSON file containing information about rmsk sequences}
}
\value{
data.table
}
\description{
Extract duplicate info from json file as a data.table
}
