% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.findResource}
\alias{.findResource}
\title{Look up the path of a resource in the cache}
\usage{
.findResource(bfc, rname)
}
\arguments{
\item{bfc}{BiocFileCache}

\item{rname}{Name of the resource}
}
\value{
Named path to the cached file (named by rid), or NULL if the
resource is not in the cache or its file does not exist
}
\description{
Look up the path of a resource in the cache
}
\keyword{internal}
