% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-through.R
\name{.fitReadThroughOnly}
\alias{.fitReadThroughOnly}
\title{Fit the read-through only model}
\usage{
.fitReadThroughOnly(y, e, dispersion, n_iter = 25L)
}
\arguments{
\item{y, e}{Matrices of observed and expected counts}

\item{dispersion}{Negative binomial dispersion}

\item{n_iter}{Number of iterations}
}
\value{
List with a, mu, and loglik
}
\description{
Fits count_ij ~ NB(a_i * expected_ij) with a_i >= 0 by iteratively
reweighted least squares.
}
\keyword{internal}
