% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.gencodeResources}
\alias{.gencodeResources}
\title{GENCODE resources for a given species}
\usage{
.gencodeResources(species)
}
\arguments{
\item{species}{Either "Hs" or "Mm"}
}
\value{
data.frame with columns type, url, file, and md5
}
\description{
GENCODE resources for a given species
}
\keyword{internal}
