% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.gencodeRname}
\alias{.gencodeRname}
\title{Name of a cached GENCODE resource}
\usage{
.gencodeRname(species, type)
}
\arguments{
\item{species}{Either "Hs" or "Mm"}

\item{type}{One of "gtf", "transcripts", or "genome"}
}
\value{
character(1)
}
\description{
Name of a cached GENCODE resource
}
\keyword{internal}
