% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.getCache}
\alias{.getCache}
\title{Get the BiocFileCache used to store rmskProfiler resources}
\usage{
.getCache(cache = NULL)
}
\arguments{
\item{cache}{NULL, a path to a cache directory, or a BiocFileCache object.
If NULL, the default rmskProfiler cache located at
\code{tools::R_user_dir("rmskProfiler", which = "cache")} is used.}
}
\value{
BiocFileCache
}
\description{
Get the BiocFileCache used to store rmskProfiler resources
}
\keyword{internal}
