% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-annotation.R
\name{.getHashOverlaps}
\alias{.getHashOverlaps}
\title{Find overlaps of TE loci regions with genomic features}
\usage{
.getHashOverlaps(x, txdb)
}
\arguments{
\item{x}{A GRanges object of TE loci}

\item{txdb}{TxDb to create annotation from}
}
\value{
List of hash vectors overlapping genomic features
}
\description{
Finds any TE that overlaps with an exon by transcript, and intron by
transcript, a 3 prime UTR by transcript, a 5 prime UTR by transcript, or
a promoter by gene.
}
