% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.getResource}
\alias{.getResource}
\title{Get the path of an existing resource in the cache}
\usage{
.getResource(bfc, rname, hint = "")
}
\arguments{
\item{bfc}{BiocFileCache}

\item{rname}{Name of the resource}

\item{hint}{Message appended to the error if the resource is missing}
}
\value{
Path to the cached file
}
\description{
Get the path of an existing resource in the cache
}
\keyword{internal}
