% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-annotation.R
\name{.getTxDb}
\alias{.getTxDb}
\title{Load the TxDb for a species from the cache, creating it if needed}
\usage{
.getTxDb(species, bfc)
}
\arguments{
\item{species}{Either "Hs" or "Mm"}

\item{bfc}{BiocFileCache}
}
\value{
TxDb
}
\description{
Load the TxDb for a species from the cache, creating it if needed
}
\keyword{internal}
