% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.META_TABLE}
\alias{.META_TABLE}
\title{Name of the BiocFileCache metadata table holding build information}
\usage{
.META_TABLE
}
\description{
Name of the BiocFileCache metadata table holding build information
}
\keyword{internal}
