% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.resourceRname}
\alias{.resourceRname}
\title{Name of a cached resource}
\usage{
.resourceRname(species, resource, exclude = NULL, min_len = NULL)
}
\arguments{
\item{species}{Either "Hs" or "Mm"}

\item{resource}{One of "gtf", "transcripts", or "genome" for GENCODE
downloads, "annotation.txdb", or one of the files in \code{.BUILD_FILES}}

\item{exclude, min_len}{Settings used to build the resource. Required for
files in \code{.BUILD_FILES}, ignored otherwise.}
}
\value{
character(1)
}
\description{
Name of a cached resource
}
\keyword{internal}
