% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.rname}
\alias{.rname}
\title{Create the name of a cached resource}
\usage{
.rname(species, file, key = NULL)
}
\arguments{
\item{species}{Either "Hs" or "Mm"}

\item{file}{Name(s) of the resource file(s)}

\item{key}{Optional settings key created by \code{.settingsKey()}}
}
\value{
character vector the same length as file
}
\description{
Resources that depend only on the species are named like
"rmskProfiler:Hs:\if{html}{\out{<file>}}". Resources that also depend on the exclude and
min_len settings include the settings key, "rmskProfiler:Hs:\if{html}{\out{<key>}}:\if{html}{\out{<file>}}".
}
\keyword{internal}
