% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.settingsKey}
\alias{.settingsKey}
\title{Create a short key identifying the settings used to build TE resources}
\usage{
.settingsKey(exclude, min_len)
}
\arguments{
\item{exclude}{Character vector of excluded RepeatMasker classes}

\item{min_len}{Minimum TE record length}
}
\value{
character(1)
}
\description{
Create a short key identifying the settings used to build TE resources
}
\keyword{internal}
