% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-resources.R
\name{.tryDownload}
\alias{.tryDownload}
\title{Attempt to download a file from the internet}
\usage{
.tryDownload(url, dest, ...)
}
\arguments{
\item{url}{URL}

\item{dest}{filepath of the destination file}

\item{...}{additional args not used}
}
\description{
Attempt to download a file from the internet
}
