% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.writeMeta}
\alias{.writeMeta}
\title{Write the rmskProfiler metadata table to the cache}
\usage{
.writeMeta(bfc, meta)
}
\arguments{
\item{bfc}{BiocFileCache}

\item{meta}{data.frame returned by \code{.readMeta()}}
}
\description{
Write the rmskProfiler metadata table to the cache
}
\keyword{internal}
