% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-resources.R
\name{downloadResources}
\alias{downloadResources}
\title{Download files needed for index generation}
\usage{
downloadResources(
  species = c("Hs", "Mm"),
  check_integrity = TRUE,
  cache = NULL
)
}
\arguments{
\item{species}{Either "Hs" (Homo sapiens) or "Mm" (Mus musculus) designating which
species to download resources for}

\item{check_integrity}{TRUE/FALSE, if TRUE check the md5sums of the GENCODE files}

\item{cache}{NULL, a path to a cache directory, or a BiocFileCache object.
Default NULL uses the rmskProfiler cache at
\code{tools::R_user_dir("rmskProfiler", which = "cache")}.}
}
\description{
This function will attempt to download all of the necessary resources for
generating the rmsk-gentrome index into the rmskProfiler cache. For humans,
it will download the GENCODE v48 transcript sequences, primary assembly, and
annotation GTF. For mouse, it will download the GENCODE M25 transcript
sequences, primary assembly, and annotation GTF. Files that already exist in
the cache are not downloaded again.
}
\examples{
\dontrun{
downloadResources(species = "Hs")
}
}
