% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-unique-seqs.R
\name{extractUniqueSeqs}
\alias{extractUniqueSeqs}
\title{Extract unique RepeatMasker sequences to a fasta file}
\usage{
extractUniqueSeqs(
  species = c("Hs", "Mm"),
  exclude = .DEFAULT_EXCLUDE,
  min_len = 32,
  cache = NULL
)
}
\arguments{
\item{species}{Either "Hs" (Homo sapiens) or "Mm" (Mus musculus)}

\item{exclude}{A character vector specifying which elements (repClass) were
excluded by \code{rmskToBed()}. Default "Simple_repeat", "Low_complexity",
"Satellite", "RNA", "rRNA", "snRNA", "scRNA", "srpRNA", "tRNA", and "Unknown".}

\item{min_len}{Minimum sequence length of a record used by \code{rmskToBed()}.
Default 32.}

\item{cache}{NULL, a path to a cache directory, or a BiocFileCache object.
Default NULL uses the rmskProfiler cache at
\code{tools::R_user_dir("rmskProfiler", which = "cache")}.}
}
\description{
This function will extract all unique RepeatMasker sequences from the genome
using the BED file generated by rmskToBed() and export these sequences to a
new fasta file in the rmskProfiler cache for downstream index generation. In
addition to producing a fasta file of the unique sequences, a json file named
rmsk-duplicateInfo.json is also saved to the cache which maps SHA1 hash values
of sequences to records in the BED file.
}
\examples{
\dontrun{
extractUniqueSeqs(species = "Hs")
}
}
