% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsk-to-bed.R
\name{rmskToBed}
\alias{rmskToBed}
\title{Extract the contents of the RepeatMasker out file into BED format}
\usage{
rmskToBed(
  species = c("Hs", "Mm"),
  exclude = .DEFAULT_EXCLUDE,
  min_len = 32,
  cache = NULL
)
}
\arguments{
\item{species}{Either "Hs" (Homo sapiens) or "Mm" (Mus musculus)}

\item{exclude}{A character vector specifying which elements to exclude from the
resulting BEd file. Default "Simple_repeat", "Low_complexity", "Satellite",
"RNA", "rRNA", "snRNA", "scRNA", "srpRNA", "tRNA", and "Unknown".}

\item{min_len}{Minimum sequence length of a record. Default 32. Records
shorter than this length are excluded from the resulting BED file.}

\item{cache}{NULL, a path to a cache directory, or a BiocFileCache object.
Default NULL uses the rmskProfiler cache at
\code{tools::R_user_dir("rmskProfiler", which = "cache")}.}
}
\description{
This function queries AnnotationHub for UCSC RepeatMasker annotations and removes all records
shorter than 31 bp, records derived from the following elements
"Simple_repeat", "Low_complexity", "Satellite", "RNA", "rRNA", "snRNA", "scRNA", "srpRNA",
"tRNA", "Unknown" and records not located in standard chromosomes. The extracted ranges are
then saved to a BED file ("rmsk.bed") in the rmskProfiler cache for downstream processing.
}
\details{
The RepeatMasker annotations used are:

AH111333 : UCSC RepeatMasker annotations (Oct2022) for Human (hg38)
AH99012 : UCSC RepeatMasker annotations (Apr2021) for Mouse (mm10)
}
\examples{
\dontrun{
rmskToBed(species = "Hs")
}

}
