% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-counts.R
\name{aggregateCounts}
\alias{aggregateCounts}
\title{Aggregate counts to gene/subfamily level}
\usage{
aggregateCounts(x, level = "subfamily")
}
\arguments{
\item{x}{SummarizedExperiment object produced by importQuants()}

\item{level}{One of "subfamily" (default), "family", or "class" indicating
the level of classification to sum TE-loci to.}
}
\value{
SummarizedExperiment
}
\description{
This function will sum counts to the gene and the subfamily (default)
level for all transcripts and TE-loci. A new SummarizedExperiment object is
returned with two assays, 'counts' and 'orig' representing the aggregated
counts from the catchSalmon down-scaled and original Salmon counts matrices,
respectively. rowData is added to the object indicating the number of TE-loci
or transcripts that were summed for each resulting feature.
}
\examples{
\dontrun{

se <- importQuants("quants", resources_dir = "hg38-resources")
aggregated <- aggregateCounts(se, level = "subfamily")

}
}
