% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-annotation.R
\name{createAnnotation}
\alias{createAnnotation}
\title{Generate annotation resources for annotating TE loci and genes}
\usage{
createAnnotation(resource_dir, keep_ranges = FALSE)
}
\arguments{
\item{resource_dir}{Path to the directory containing index generation resources.
Output is saved to this location.}

\item{keep_ranges}{Should a GRangesList of each transcript/TE locus be saved in the annotation
object? default FALSE}
}
\description{
This function produces multiple annotation files used to annotate downstream
counts generated after quantification with Salmon. It creates a GRangesList
object for each hash and a tab-delimited file containing annotation
information for each hash with indicator values for the types of genic
features they overlap. Overlap annotations are generated with respect to the
downloaded gencode GTF file.
}
\examples{
\dontrun{
createAnnotation(resource_dir = "/path/to/rmsk-resources")
}
}
