% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-annotation.R
\name{createAnnotation}
\alias{createAnnotation}
\title{Generate annotation resources for annotating TE loci and genes}
\usage{
createAnnotation(resource_dir, stranded = TRUE)
}
\arguments{
\item{resource_dir}{Path to the directory containing index generation resources.
Output is saved to this location.}

\item{stranded}{Should the annotation be created with respect to strand? Default
TRUE. If stranded=FALSE then overlap annotation is performed on sequence
positions without respect to strand.}
}
\description{
This function produces multiple annotation files used to annotate downstream
counts generated after quantification with Salmon. It creates: GRangesList
object for each Hash value and all unique ranges for the hash and A
tab-delimited file of unique hash-by-overlap information. This function
generates overlap annotations with respect to the annotations provided by
annotatr::build_annotations() on either hg38_basicgenes or mm10_basicgenes,
depending on the resources detected.
}
\examples{
\dontrun{
createAnnotation(resource_dir = "/path/to/rmsk-resources")
}
}
