% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-gentrome.R
\name{createGentrome}
\alias{createGentrome}
\title{Create the gentrome fasta and Salmon index}
\usage{
createGentrome(resource_dir, create_index = TRUE, threads = 1)
}
\arguments{
\item{resource_dir}{Path to the directory containing index generation resources.
Output is saved to this location.}

\item{create_index}{Create salmon index after generating resources? Default TRUE.
This assumes that "salmon" is available on your PATH}

\item{threads}{Number of threads to use for salmon index generation. Default 1}
}
\description{
This function creates the gentrome (transcripts + unique rmsk sequences +
decoy) fasta and (optionally) the Salmon index. If index generation is
desired then the function assumes that Salmon is on your PATH. This function
will create a gentrome.fa file, decoys.txt file, and (optionally) a Salmon
index from these files in the resource directory.
}
\examples{
\dontrun{
createGentrome(resource_dir = "/path/to/rmsk-resources")
}
}
