% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch-salmon2.R
\name{.catchSalmon2}
\alias{.catchSalmon2}
\title{Import transcript-level counts with offset}
\usage{
.catchSalmon2(paths, verbose = TRUE)
}
\arguments{
\item{character}{vector giving paths to the sample-specific directories created by a kallisto or
Salmon. Each entry corresponds to one RNA-seq sample.}

\item{logical}{If TRUE, progress information will be sent to standard output as each sample is processed.}
}
\description{
Import transcript-level counts with offset
}
\details{
This function is mostly lifted from:

https://github.com/plbaldoni/TranscriptDE-code/blob/main/code/pkg/R/utils.R#L52

and combined with existing catchSalmon code. I only used data.table::fread in place of readr so
as to not add more dependencies.
}
\keyword{internal}
