% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.newResource}
\alias{.newResource}
\title{Get a path in the cache to write a generated resource to}
\usage{
.newResource(bfc, rname, ext)
}
\arguments{
\item{bfc}{BiocFileCache}

\item{rname}{Name of the resource}

\item{ext}{File extension of the resource, e.g. ".bed"}
}
\value{
Path to the cached file
}
\description{
If the resource already exists in the cache its path is returned so that it
is overwritten. Otherwise a new cache entry is created.
}
\keyword{internal}
