% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-unique-seqs.R
\name{extractUniqueSeqs}
\alias{extractUniqueSeqs}
\title{Extract unique RepeatMasker sequences to a fasta file}
\usage{
extractUniqueSeqs(resource_dir, remove_tmp = TRUE)
}
\arguments{
\item{resource_dir}{Path to the directory containing index generation resources.
Output is saved to this location.}

\item{remove_tmp}{The genome fasta file needs to be temporarily unzipped for
feature extraction. Should the unzipped file be removed after processing?
Default TRUE}
}
\description{
This function will extract all unique RepeatMasker sequences from the genome
using the BED file generated by rmskToBed() and export these sequences to a
new fasta file for downstream index generation. In addition to producing a
fasta file of the unique sequences, a json file is also dumped to the
resource directory named rmsk-duplicateInfo.json which maps SHA1 hash values
of sequences to records in the BED file.
}
\examples{
\dontrun{
extractUniqueSeqs(resource_dir = "/path/to/rmsk-resources")
}
}
