% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-index.R
\name{generateIndex}
\alias{generateIndex}
\title{Generate a Salmon index and annotations for TE quantification}
\usage{
generateIndex(
  out_dir,
  species = c("Hs", "Mm"),
  check_integrity = TRUE,
  exclude = c("Simple_repeat", "Low_complexity", "Satellite", "RNA", "rRNA", "snRNA",
    "scRNA", "srpRNA", "tRNA", "Unknown"),
  min_len = 32,
  keep_ranges = TRUE,
  create_index = TRUE,
  threads = 1
)
}
\arguments{
\item{out_dir}{Directory in which to save all resources including final generated Salmon index}

\item{species}{Either "Hs" (Homo sapiens) or "Mm" (Mus musculus) designating which
species to download resources for}

\item{check_integrity}{TRUE/FALSE, if TRUE check the md5sums of the GENCODE files after downloading}

\item{exclude}{A character vector specifying which elements (repClass) to exclude from the
resulting BEd file. Default "Simple_repeat", "Low_complexity", "Satellite",
"RNA", "rRNA", "snRNA", "scRNA", "srpRNA", "tRNA", and "Unknown".}

\item{min_len}{Minimum sequence length of a record. Default 32. Records
shorter than this length are excluded from the resulting BED file and index.}

\item{keep_ranges}{Should a GRangesList of each transcript/TE locus be saved in the annotation
object? default TRUE}

\item{create_index}{TRUE/FALSE Create salmon index after generating resources? Default TRUE.
This assumes that "salmon" is available on your PATH}

\item{threads}{Number of threads to use for salmon index generation. Default 1}
}
\description{
This function is a wrapper around several other functions for downloading and
generating all of the files needed to create a Salmon index in order to
quantify transcripts and TEs at the loci level.
}
\details{
This function creates a Salmon index from all GENCODE transcripts + selected
RepeatMasker elements + a genomic decoy. It does so by calling the following
functions:

\itemize{
\item{"downloadResources()"}{Downloads all neccessary files for downstream processing}
\item{"rmskToBed()"}{Extracts and filters RepeatMasker to BED file}
\item{"extractUniqueSeqs()"}{Extracts all unique RepeatMasker sequences from the genome}
\item{"createAnnotation()"}{Creates annotations of all TE loci with respect to annotations in GTF}
\item{"createGentrome()"}{Creates the files needed for decoy-aware Salmon index and calls 'Salmon index'}
}
}
\examples{
\dontrun{
# Run pipeline for downloading and creating Salmon index and annotations
generateIndex(out_dir = "rmsk-resources", species = "hs")
}
}
