% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsk-to-bed.R
\name{rmskToBed}
\alias{rmskToBed}
\title{Extract the contents of the RepeatMasker out file into BED format}
\usage{
rmskToBed(
  resource_dir,
  exclude = c("Simple_repeat", "Low_complexity", "Satellite", "RNA", "rRNA", "snRNA",
    "scRNA", "srpRNA", "tRNA", "Unknown"),
  min_len = 32
)
}
\arguments{
\item{resource_dir}{Path to the directory where gentrome resources were
downloaded. this should be the same path specified by downloadResources().}

\item{exclude}{A character vector specifying which elements to exclude from the
resulting BEd file. Default "Simple_repeat", "Low_complexity", "Satellite",
"RNA", "rRNA", "snRNA", "scRNA", "srpRNA", "tRNA", and "Unknown".}

\item{min_len}{Minimum sequence length of a record. Default 32. Records
shorter than this length are excluded from the resulting BED file.}
}
\description{
This function extracts the contents of the RepeatMasker out file into a BED
file. Existing BED files for these tracks exist however this function allows
for arbitrary filtering out of certain records by feature name or feature
length. The default behavior removes all records shorter than 31 bp and
records derived from the following elements "Simple_repeat", "Low_complexity",
"Satellite", "RNA", "rRNA", "snRNA", "scRNA", "srpRNA", "tRNA", "Unknown".
}
\examples{
\dontrun{
rmskToBed(resource_dir = "/path/to/rmsk-resources")
}
}
