% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon-quant.R
\name{salmonQuant}
\alias{salmonQuant}
\title{Quantify TEs and transcripts using Salmon with Gibbs sampling}
\usage{
salmonQuant(fq1, fq2, sample_names, resource_dir, out_dir, nGibbs = 30, ...)
}
\arguments{
\item{fq1}{vector of file paths to fastq read 1 files}

\item{fq2}{vector of file paths to fastq read 2 files}

\item{sample_names}{character vector of samples names matching each pair of fastq files}

\item{resource_dir}{path to the rmskProfiler resource directory containing an rmsk.salmon_index directory}

\item{out_dir}{path to save the quant directories for each sample. This will
be the parent directory, samples are saved in subdirectories like out_dir/<sample_name>_quants}

\item{nGibbs}{integer number of Gibbs samples to perform. Default 30. Published
work suggests this number can be set to around 180 / length(samples) but 30
ensures a good coverage with minimal cost in terms of speed.}

\item{...}{Additional arguments passed to Salmon as character strings, e.g.
"--gcBias", "--seqBias" "--threads 12", see examples below}
}
\description{
This function provides a simple wrapper around a system call to
\code{salmon quant ...}. The default arguments assume that you have paired
end reads (which you should if quantifying TEs is a priority). The defaults
DO NOT set any other flags for Salmon other than automatically detecting the
library type and setting the number of Gibbs samples. You can (and should)
pass additional arguments as character strings with the desired Salmon flags.
See the example below for how to do this.
}
\examples{
\dontrun{

# Specify paths to fastq files and names of the samples
fq1 <- c("sample1.R1.fq.gz", "sample2.R1.fq.gz", "sample3.R1.fq.gz")
fq2 <- c("sample1.R2.fq.gz", "sample2.R2.fq.gz", "sample3.R2.fq.gz")
sample_names <- c("sample1", "sample2", "sample3")

# Perform with all default settings -- probably not recommended
salmonQuant(fq1, fq2, sample_names, resource_dir = "hg38-resources", out_dir = "quants")

# More often though we will want to pass additional arguments to Salmon
# We can do so by providing additoinal flags as a character strings
salmonQuant(
  fq1 = fq1,
  fq2 = fq2,
  sample_names = sample_names,
  resource_dir = "hg38-resources",
  out_dir = "quants",
  "--gcBias",
  "--seqBias",
  "--posBias",
  "--threads 12"
  )
}
}
